+++ src/components/NoticesFilters/NoticesFilter.styled.jsx
@@ -1,8 +1,12 @@
 import styled from "styled-components";
 import { ReactComponent as Filter } from '../../images/icons/filters-3.svg';
 
+export const FilterWrapper = styled.div`
+  position: relative;
+`
+
 export const Button = styled.button`
-  color: #54ADFF;
+  color: var(--cl-blue-link);
   font-size: 16px;
   font-weight: 700;
   width: 152px;
@@ -11,14 +15,14 @@ export const Button = styled.button`
   display: flex;
   align-items: center;
   justify-content: center;
-  background-color: #fff;
+  background-color: var(--cl-almost-white);
   border: 2px solid rgb(84, 173, 255);
   border-radius: 40px;
   transition: background-color 250ms ease-in-out 0s, color 250ms ease-in-out 0s;
   cursor: pointer;
 
   &:hover {
-    background: linear-gradient(290.46deg, #419EF1 0%, #9BD0FF 107.89%);
+    background: var(--gr-blue);
     color: var(--cl-background);
     border: none;
 
@@ -29,6 +33,8 @@ export const Button = styled.button`
 `
 
 export const FilterCSS = styled(Filter)`
+  wirth: 21px;
+  height: 24px;
   margin-left: 12px;
   stroke: currentColor;
 `
@@ -37,8 +43,8 @@ export const FilterSelect = styled.div`
   width: 152px;
   padding: 8px;
   background-color: var(--cl-background);
-  color: #54ADFF;
-  box-shadow: 3px 8px 14px rgba(136, 198, 253, 0.19);
+  color: var(--cl-blue-link);
+  box-shadow: var(--shadow-default);
   border-radius: 20px;
   position: absolute;
   z-index: 100;
@@ -52,8 +58,8 @@ export const FilterSelectTitle = styled.h3`
 `
 
 export const FiltersButton = styled.button`
-  padding: 4px 8px;
-  background: #CCE4FB;
+  padding: 12px 8px;
+  background: var(--cl-blue-light);
   border-radius: 20px;
   border: none;
   outline: none;
@@ -71,7 +77,7 @@ export const FiltersButton = styled.button`
   p {
     font-size: 14px;
     margin-left: 14px;
-    color: #54ADFF;
+    color: var(--cl-blue-link);
   }
 
   input {
@@ -80,7 +86,7 @@ export const FiltersButton = styled.button`
   }
 
   label {
-    color: #54ADFF;
+    color: var(--cl-blue-link);
     font-size: 12px;
     display: flex;
     align-items: center;
@@ -95,4 +101,18 @@ export const FiltersButton = styled.button`
       margin-right: 16px;
     }
   }
+`
+
+export const FiltersChevron = styled.svg`
+ width: 24px;
+ height: 24px;
+ fill: var(--cl-blue-light);
+ stroke: var(--cl-blue-link);
+`
+
+export const FiltersRound = styled.svg`
+ width: 20px;
+height: 20px;
+fill: var(--cl-blue-light);
+stroke: var(--cl-blue-link);
 `
