+++ src/components/NoticesPage/NoticesPage.styled.jsx
@@ -9,56 +9,32 @@ export const ModalContent = styled.div`
   padding-bottom: 60px;
 
   font-style: normal;
-
+  letter-spacing: -0.01em;
   color: var(--cl-black);
 `;
 
 export const TitleModal = styled.h4`
-  width: 240px;
-  
-  font-size: 24px;
-  line-height: 33px;
-  text-align: center;
-  letter-spacing: 0.04em;
+  font-weight: 700;
+  font-size: 28px;
+  line-height: 38px;
 
-  padding-bottom: 14px;
-
-  @media (min-width: 768px) {
-    width: 317px;
-    font-size: 28px;
-    line-height: 38px;
-    padding-bottom: 40px;
-  }
+  padding: 0;
+  margin-top: 0px;
+  margin-bottom: 40px;
 `;
 
-export const DescrModal = styled.div`
+export const DescrModal = styled.p`
   font-weight: 500;
   font-size: 16px;
   line-height: 22px;
   text-align: center;
-
-  margin-bottom: 48px;
-`;
-
-export const Text = styled.p`
-  width: 240px;
-
-  @media (min-width: 768px) {
-    width: 390px;
-  }
 `;
 
 export const ButtonWrap = styled.div`
   display: flex;
-  flex-direction: column;
   justify-content: center;
   align-items: center;
-  gap: 5px;
-
-  @media (min-width: 768px) {
-    flex-direction: row;
-    gap: 17px;
-  }
+  gap: 17px;
 `;
 
 export const Button = styled.button`
@@ -67,33 +43,24 @@ export const Button = styled.button`
   justify-content: center;
   align-items: center;
   gap: 8px;
-
   padding: 0;
-
-  font-weight: 700;
-  font-size: 16px;
-  line-height: 22px;
-
+  width: 129px;
+  height: 40px;
   letter-spacing: 0.04em;
   font-weight: 700;
 
   border: 2px solid var(--cl-blue-link);
   border-radius: 40px;
   color: var(--cl-blue-link);
-  background-color: var(--cl-white);
-  width: 256px;
-  height: 40px;
-  @media (min-width: 768px) {
-    width: 129px;
-  }
+  background-color: var(--cl-background);
 
-  transition: background-color var(--animat), color var(--animat);
+  transition: background-color var(--animat), color var(--animat),
+    border-color var(--animat);
 
-  &:hover,
-  &:active,
-  &:focus {
+  &:hover {
     background-color: var(--cl-blue-link);
-    color: var(--cl-white);
+    color: var(--cl-background);
+    border-color: var(--cl-background);
   }
 `;
 
@@ -103,3 +70,9 @@ export const Trash = styled.svg`
   stroke: currentColor;
   fill: transparent;
 `;
+
+export const Filters = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+`
