--- src/pages/Notices.styled.jsx
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-import '../index.css'
-
-export const Filters = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`
